edited index.css to not put plant image on all bodies
Sohpia - if you see this I had to slightly change the index.css that so that your plant image did not affect 'body' because 'body' affects all pages. I'll explain in facebook, but you still may want to adjust how the plant image shows on home page. Have a look.